MaxDaysPerWeek fix
Adapted TeacherService to take the minimum of the given maxDaysPerWeek and the amount of days the teacher is actually available each week. E.G. A teacher that claims to be available 4 days a week but is only available on friday will be stored as max 1 day per week. A teacher available every weekday but only 2 days per week will be saved as max 2 days per week.