Access point add teacher
Add Teacher button added to View Teachers screen. Clicking this button shows a modal for inputting a name, teaches practicum, availability for every workday and for the maximum amount of days per week. This data is passed on and handled by the addTeacher endpoint to add the Teacher to the database.