Override by dragging
The admin can now drag every lesson to a new location in the table. Doing so opens a modal which takes the date and classroom where the lesson was dropped. The admin can now submit this change. The admin has the option of amending the new date and classroom from the modal as well, and to fill in a checkbox allowing the change to be made weekly. Submitting this change will call on the override endpoint which changes the selected lesson's date and classroom in the database.